requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
snowflake-connector-python==3.12.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.2
